Changed licensing configuration
Since all the sources are generated, it makes no sense to remove the
license post-packaging. Generated source files are now annotated with
the license, which itself is preceded by a warning to not edit the files
manually. This warning is also present in the test files. The license is
not included in those because we do not distribute them.